WEB 20160730 02 bruno - [bug fix] sometime we could not create a symlink on prod server - Increase timeout to allow more PHP calculation time - [bug fix] make sure a temp project exists

+++ libs/Folders.php
@@ -56,7 +56,7 @@ class Folders {
 			}
 		}
 		return true;
-	}	
+	}
 
 	public function loopFolder(){
 		$list = array();
@@ -93,10 +93,12 @@ class Folders {
 	}
 
 	public function createSymlink($target, $link){
-		if($this->createPath($link)){
-			if(rmdir($link)){
-				if(symlink($target, $link)){
-					return $this->setPath($link);
+		if(!is_dir($link)){
+			if($this->createPath($link)){
+				if(rmdir($link)){
+					if(symlink($target, $link)){
+						return $this->setPath($link);
+					}
 				}
 			}
 		}
